+++ backend/app/api/users/manager.js
@@ -1,6 +1,6 @@
 const AccessRestriction = require("../../models/access-restriction.model");
 const { User } = require("../../models");
-const { storeFile, readFile, deleteFile } = require("../../utils/file");
+const { storeFile, readFile, deleteFile, validateAndNormaliseFile } = require("../../utils/file");
 
 const picture = (userId) => `users/${userId}/picture`;
 
@@ -11,16 +11,17 @@ const picture = (userId) => `users/${userId}/picture`;
 function createUser(user) {
     const { pictureUrl, ...pureUser } = user;
     pureUser.access = AccessRestriction.user;
+    const pict = validateAndNormaliseFile(pictureUrl, /image\/.*/);
     let result = User.create(pureUser);
-    User.update(result.id, { pictureUrl: storeFile(picture(result.id), pictureUrl) });
+    User.update(result.id, { pictureUrl: storeFile(picture(result.id), pict) });
     return { ...result, pictureUrl };
 }
 
 function replaceUser(userId, user) {
     const { pictureUrl, ...pureUser } = user;
     const previousUser = User.getById(userId);
-    deleteFile(previousUser.pictureUrl);
-    pureUser.pictureUrl = storeFile(picture(userId), pictureUrl);
+    pureUser.pictureUrl = storeFile(picture(userId), ...validateAndNormaliseFile(pictureUrl, /image\/.*/));
+    if (pureUser.pictureUrl && previousUser.pictureUrl !== pureUser.pictureUrl) deleteFile(previousUser.pictureUrl);
     const result = User.replace(userId, pureUser);
     return { ...result, pictureUrl: readFile(result.pictureUrl) };
 }
@@ -28,8 +29,8 @@ function replaceUser(userId, user) {
 function updateUser(userId, user) {
     const { pictureUrl, ...pureUser } = user;
     const previousUser = User.getById(userId);
-    deleteFile(previousUser.pictureUrl);
-    pureUser.pictureUrl = storeFile(picture(userId), pictureUrl);
+    pureUser.pictureUrl = storeFile(picture(userId), ...validateAndNormaliseFile(pictureUrl, /image\/.*/));
+    if (pureUser.pictureUrl && previousUser.pictureUrl !== pureUser.pictureUrl) deleteFile(previousUser.pictureUrl);
     const result = User.update(userId, pureUser);
     return { ...result, pictureUrl: readFile(result.pictureUrl) };
 }
